Let the event-type pool grow past 255
It was the one ceiling a real device might meet - the only table that accumulates
across channels, so thirteen channels averaging twenty types each is already 260 -
and the one that cost nothing to lift. Nothing addresses the pool: a type's wire
index is counted within its channel, which _findEventType does by walking the pool
and counting only that channel's entries. So the capacity, the slot count and the
running count widen to uint16_t, the clamp goes, and the protocol is untouched.

Five loops over the pool widened with them. That part is not optional: a byte
counter against a limit above 255 wraps to zero and never terminates, so widening
a capacity without its loops turns a full table into a hang. There are 28 such
loops across the tables; the five over the event-type pool are the ones this
commit touches, and none over _eventTypeCount remain.

Costs two bytes of SRAM and about 150 of flash on AVR, for 16-bit arithmetic on
an 8-bit machine. EventChannels goes 30208 -> 30360 flash, 894 -> 896 SRAM.

Commands are in the same position - matched by the hash of the name, no index on
the wire - and left alone for now: 255 commands on one device is not a shape
anyone builds, and it is nine more loops and four signatures for it, each a chance
to leave a byte counter behind. State and event channels genuinely cannot move
without changing 0x95 and 0x85.

+++ src/BlaeckSerial.h
@@ -454,19 +454,22 @@ class BlaeckSerial;
 // builds and the number is simply not stored, so a feature switch never breaks
 // a chain.
 //
-// Everything but the signals caps at 255 entries, and asking for more is clamped
-// there with a line on the debug stream. Where that 255 comes from differs:
+// Three of these cap at 255 entries, and asking for more is clamped there with a
+// line on the debug stream. Which three, and why, is worth knowing before sizing a
+// large device:
 //
 //   withStateChannels   the wire. A value names its channel with a one-byte index,
 //   withEventChannels   so a 256th channel could be declared and never addressed.
 //
-//   withEventTypes      this library's own counters. A type's index is counted
-//   withCommands        within its channel rather than across the pool, and a
-//                       command is matched by the hash of its name, so neither
-//                       has a wire index that could run out.
+//   withCommands        this library's own counter. A command is matched by the
+//                       hash of its name and carries no index on the wire, so this
+//                       255 could be lifted without the protocol noticing.
 //
 //   withSignals         nothing. A data frame carries values in catalog order and
 //                       names none of them, so there is no index at all.
+//
+//   withEventTypes      nothing. A type's index is counted within its channel, so
+//                       the pool is addressed by nothing and holds what RAM allows.
 //
 // Long before any of that, RAM decides. A state entry is around 25 bytes, so 255 of
 // them is most of a Mega's SRAM; on an Uno the arithmetic never gets that far.
@@ -498,10 +501,10 @@ public:
   // channel: they share one table, each entry tagged with the channel that
   // declared it. Four channels of five types each need withEventTypes(20).
   //
-  // The most reachable of these ceilings, being the only one that accumulates -
-  // thirteen channels averaging twenty types each would find it. On the wire a
-  // type's index is counted within its channel, so 255 is this library's limit
-  // rather than the protocol's.
+  // The only one of these tables with no ceiling but RAM. It is also the only one
+  // that accumulates - thirteen channels averaging twenty types each is 260 - and
+  // nothing addresses the pool itself, since a type's wire index is counted within
+  // its channel. Costs two bytes of SRAM over a byte-wide count.
   BlaeckBeginRef &withEventTypes(unsigned int count);
 
   // Room for the commands registered with onCommand() and with the typed
@@ -2939,10 +2942,10 @@ private:
   // beside the other _buf helpers because it needs EventTypeEntry, declared just above.
   void _bufEventType0(const EventTypeEntry &e);
   EventTypeEntry *_eventTypes = nullptr;
-  byte _eventTypeCapacity = DEFAULT_EVENT_TYPES;
-  byte _eventTypeSlots() const { return _eventTypes != nullptr ? _eventTypeCapacity : 0; }
+  uint16_t _eventTypeCapacity = DEFAULT_EVENT_TYPES;
+  uint16_t _eventTypeSlots() const { return _eventTypes != nullptr ? _eventTypeCapacity : 0; }
   bool _ensureEventTypeTable();
-  byte _eventTypeCount = 0;
+  uint16_t _eventTypeCount = 0;
 #endif
   BlaeckAnyCommandHandler _anyCommandHandler = nullptr;
   char _parsedTokenBuffer[MAXIMUM_CHAR_COUNT] = {0};
